+++ Gruntfile.js
@@ -1,16 +1,12 @@
-'use strict';
+'use strict'
 
 module.exports = function (grunt) {
-    require('load-grunt-tasks')(grunt);
+    grunt.loadNpmTasks('grunt-babel')
+    grunt.loadNpmTasks('grunt-contrib-clean')
+    grunt.loadNpmTasks('grunt-mocha-test')
 
     grunt.initConfig({
         babel: {
-            options: {
-                sourceMaps: 'inline',
-                nonStandard: false,
-                optional: [ 'runtime', 'strict' ],
-                stage: 0,
-            },
             src: {
                 expand: true,
                 src: 'src/**/*.js',
@@ -29,11 +25,11 @@ module.exports = function (grunt) {
         mochaTest: {
             src: 'target/test/src/**/*.js'
         }
-    });
+    })
 
-    grunt.registerTask('compile:src', [ 'babel:src' ]);
-    grunt.registerTask('compile:test', [ 'babel:test' ]);
-    grunt.registerTask('compile', [ 'compile:src', 'compile:test' ]);
-    grunt.registerTask('test', [ 'compile', 'mochaTest' ]);
-    grunt.registerTask('default', [ 'test' ]);
-};
+    grunt.registerTask('compile:src', [ 'babel:src' ])
+    grunt.registerTask('compile:test', [ 'babel:test' ])
+    grunt.registerTask('compile', [ 'compile:src', 'compile:test' ])
+    grunt.registerTask('test', [ 'compile', 'mochaTest' ])
+    grunt.registerTask('default', [ 'test' ])
+}